requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
drf-yasg>=1.21.0
Pillow>=9.5.0  # for image handling
requests>=2.28.0  # for making HTTP requests
python-dotenv>=1.0.0  # for environment variables